requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
class-registry==2.1.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.2
django-filter==23.2
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.5
filters==1.3.2
filters-django==1.0.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
openapi-codec==1.3.2
packaging==23.1
psycopg2==2.9.6
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
regex==2023.3.23
requests==2.29.0
ruamel.yaml==0.17.22
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==1.26.15
textract~=1.6.5